Rename urlfieldinverter* -> url_field_inverter*.
