+++ searchlib/src/tests/memoryindex/url_field_inverter/.gitignore
@@ -1,0 +1,1 @@
+searchlib_url_field_inverter_test_app
