+++ searchlib/src/tests/memoryindex/url_field_inverter/CMakeLists.txt
@@ -1,0 +1,9 @@
+# Copyright 2017 Yahoo Holdings. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
+vespa_add_executable(searchlib_url_field_inverter_test_app TEST
+    SOURCES
+    url_field_inverter_test.cpp
+    DEPENDS
+    searchlib_test
+    searchlib
+)
+vespa_add_test(NAME searchlib_url_field_inverter_test_app COMMAND searchlib_url_field_inverter_test_app)
