--- searchlib/src/tests/memoryindex/urlfieldinverter/.gitignore
@@ -1,1 +1,0 @@
-searchlib_urlfieldinverter_test_app
